Level 500: Refactor the code
