+++ index.html
@@ -9,6 +9,14 @@
     <title>TV Show Project | Ali Qassab</title>
     <link href="style.css" rel="stylesheet" />
     <link rel="icon" href="/favicon.ico" type="image/x-icon" />
+
+    <!-- Preconnect to TVMaze domains -->
+    <link rel="preconnect" href="https://api.tvmaze.com" />
+    <link rel="preconnect" href="https://static.tvmaze.com" />
+
+    <!-- Fallback DNS prefetch -->
+    <link rel="dns-prefetch" href="https://api.tvmaze.com" />
+    <link rel="dns-prefetch" href="https://static.tvmaze.com" />
   </head>
 
   <body>
